feat(migration): all models in one file
**What**
- relocated database models
- delete changes in migrations.env
